refactor: move all config items to ConfigUtils class
