+++ src/main/java/org/opentripplanner/middleware/OtpMiddlewareMain.java
@@ -1,7 +1,6 @@
 package org.opentripplanner.middleware;
 
 import com.beerboy.ss.SparkSwagger;
-import com.fasterxml.jackson.databind.JsonNode;
 import org.eclipse.jetty.http.HttpStatus;
 import org.opentripplanner.middleware.auth.Auth0Connection;
 import org.opentripplanner.middleware.bugsnag.BugsnagJobs;
@@ -15,18 +14,16 @@ import org.opentripplanner.middleware.controllers.api.MonitoredTripController;
 import org.opentripplanner.middleware.controllers.api.TripHistoryController;
 import org.opentripplanner.middleware.otp.OtpRequestProcessor;
 import org.opentripplanner.middleware.persistence.Persistence;
+import org.opentripplanner.middleware.utils.ConfigUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import spark.Service;
 
-import java.io.File;
-import java.io.FileInputStream;
 import java.io.IOException;
 import java.util.List;
 
 import static javax.ws.rs.core.MediaType.APPLICATION_JSON;
 import static org.opentripplanner.middleware.utils.JsonUtils.logMessageAndHalt;
-import static org.opentripplanner.middleware.utils.YamlUtils.yamlMapper;
 
 /**
  * Main class for OTP Middleware application. This handles loading the configuration files, initializing the MongoDB
@@ -34,14 +31,12 @@ import static org.opentripplanner.middleware.utils.YamlUtils.yamlMapper;
  */
 public class OtpMiddlewareMain {
     private static final Logger LOG = LoggerFactory.getLogger(OtpMiddlewareMain.class);
-    private static final String DEFAULT_ENV = "configurations/default/env.yml";
-    private static JsonNode envConfig;
     private static final String API_PREFIX = "/api/";
     public static boolean inTestEnvironment = false;
 
     public static void main(String[] args) throws IOException {
         // Load configuration.
-        loadConfig(args);
+        ConfigUtils.loadConfig(args);
 
         // Connect to MongoDB.
         Persistence.initialize();
@@ -115,104 +110,4 @@ public class OtpMiddlewareMain {
             return null;
         });
     }
-
-    /**
-     * Load config files from either program arguments or (if no args specified) from
-     * default configuration file locations. Config fields are retrieved with getConfigProperty.
-     */
-    private static void loadConfig(String[] args) throws IOException {
-        FileInputStream envConfigStream;
-
-        if (args.length == 0) {
-            LOG.warn("Using default env.yml: {}", DEFAULT_ENV);
-            envConfigStream = new FileInputStream(new File(DEFAULT_ENV));
-        } else {
-            LOG.info("Loading env.yml: {}", args[0]);
-            envConfigStream = new FileInputStream(new File(args[0]));
-        }
-
-        envConfig = yamlMapper.readTree(envConfigStream);
-    }
-
-    /**
-     * Convenience function to get a config property (nested fields defined by dot notation "data.use_s3_storage") as
-     * JsonNode. Checks env.yml and returns null if property is not found.
-     */
-    private static JsonNode getConfigProperty(String name) {
-        String[] parts = name.split("\\.");
-        JsonNode node = envConfig;
-        for (String part : parts) {
-            if (node == null) {
-                LOG.warn("Config property {} not found", name);
-                return null;
-            }
-            node = node.get(part);
-        }
-        return node;
-    }
-
-    /**
-     * Convenience function to check existence of a config property (nested fields defined by dot notation
-     * "data.use_s3_storage") in either server.yml or env.yml.
-     */
-    public static boolean hasConfigProperty(String name) {
-        // try the server config first, then the main config
-        return hasConfigProperty(envConfig, name);
-    }
-
-    private static boolean hasConfigProperty(JsonNode config, String name) {
-        String[] parts = name.split("\\.");
-        JsonNode node = config;
-        for (String part : parts) {
-            if (node == null) {
-                return false;
-            }
-            node = node.get(part);
-        }
-        return node != null;
-    }
-
-    /**
-     * Get a config property (nested fields defined by dot notation "data.use_s3_storage") as text.
-     */
-    public static String getConfigPropertyAsText(String name) {
-        JsonNode node = getConfigProperty(name);
-        if (node != null) {
-            return node.asText();
-        } else {
-            LOG.warn("Config property {} not found", name);
-            return null;
-        }
-    }
-
-    /**
-     * @return a config value (nested fields defined by dot notation "data.use_s3_storage") as text or the default value
-     * if the config value is not defined (null).
-     */
-    public static String getConfigPropertyAsText(String name, String defaultValue) {
-        JsonNode node = getConfigProperty(name);
-        if (node != null) {
-            return node.asText();
-        } else {
-            return defaultValue;
-        }
-    }
-
-    /**
-     * @return a config value (nested fields defined by dot notation "data.use_s3_storage") as an int or the default
-     * value if the config value is not defined (null) or cannot be converted to an int.
-     */
-    public static int getConfigPropertyAsInt(String name, int defaultValue) {
-
-        int value = defaultValue;
-
-        try {
-            JsonNode node = getConfigProperty(name);
-            value = Integer.parseInt(node.asText());
-        } catch (NumberFormatException | NullPointerException e) {
-            LOG.warn("Unable to parse {}. Using default: {}", name, defaultValue, e);
-        }
-        return value;
-    }
-
 }
