+++ src/main/java/org/opentripplanner/middleware/bugsnag/BugsnagReporter.java
@@ -6,7 +6,7 @@ import com.bugsnag.Severity;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import static org.opentripplanner.middleware.OtpMiddlewareMain.getConfigPropertyAsText;
+import static org.opentripplanner.middleware.utils.ConfigUtils.getConfigPropertyAsText;
 
 /**
  * Bugsnag util for reporting errors to the project defined by the Bugsnag project notifier API key.
